latest changes from worth serious reviewing
